Stopping JMS Embedded Server via CLosing View Function. So click FILE -> CLOSE to safely shut down for JMS.
Changed some minor things in Producer so it is being produced, works
now, it should be in Queue. Next is to write a MessageListener who waits
for Messages and handles them and creates Jobs for Painter and Montage
Worker.

added to Build.xml run-jms-client, but not working since it does not
find hornetq-configuration.xml although config file for jms is provided
via classpath?