Added JobManagementListener, with functionality.
Not working at the moment, since it does not listen on the queue and
calls up onMessage (Possible Issues: Not writing into queue at all, or
listening to wrong queue, or added listener in a wrong way, somehow the
liustener has to start listening and knowing of queue etc... therfore i
startet it now in the Ctor within CarFactoryManager.java so it does not
get lost). Still not working.

Creating Assemble and painterJob should be working at least the
functionality itself.

created jms and xvsm worker package since a little bite different.