JMS MessageListener is working now correctly.
If you sstart CarFactory for JMS then with log-outputs it shoudl be
visible that AssembleJob and PainterJob creation and message sending out
to respective queues is working.

Create 8 Tires, 2 Motoren, then 1 Body, as soon as the Body is finished
it should detected that an assembleJob is available see Logs.
Create another CarBody and the next assembleJob is created, and so on...

In WorkTask I added a atomicLong ID so each WorkTask has an ID, as soon
as more than one carPart is generated this ID is incremented and set to
the new value so it represents the carPartId correctly, see Producer and
WorkTask (JMS).

Small fixes due to testing.